+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Add refresh token to Dashboard
 + Add single fetch. Prepare for call detail
++ Add open close modal
 
 ## 0.16.0
 + Add dashboard